python-panignation/requirements.txt
Werkzeug==1.0.1
psycopg2==2.8.5
Flask==1.1.2
Flask_Cors==3.0.10
python-dotenv==0.18.0
